requirements.txt
discord.py
python-dateutil
pandas
arrow
pendulum
dateparser
